Fix duplicate NSManagedObjects when processing version data.
Assume `objectKeys` contained a value x and a value y such that x == y
and there is no managed object with simperium key `x`. Then when
iterating over versions, multiple coredata objects with the same
simperium key will be created.

The fix here is to create the object once and update the objects array
so that the created object may be reused if needed.